Animated piece rotation
I also moved piece dimming (for when the piece has been set to the correct position) to inside the Rotate coroutine (rather than called through SetPiece in the PuzzlePieceManager class), so that it does not dim before the rotation animation completes. There is no animation when the puzzle pieces are shuffled when the puzzle loads.