Sped up puzzle piece UI fade in when new puzzle loads
Ocassionally, toward the beginning of assembling a puzzle, pieces would become 100% alpha after locking them in place (when they should be a lower alpha to show the locked state). The fade in animation was taking too long and overlapping user interaction, overriding the lower alpha setting and making the piece appear that it was not yet placed.